Rename differ to differ-cli and instantiate a core module
