+++ tools/differ-cli/src/commonMain/kotlin/io/matthewnelson/differ/cli/internal/ArgTypePath.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+package io.matthewnelson.differ.cli.internal
+
+import kotlinx.cli.ArgType
+import okio.Path
+import okio.Path.Companion.toPath
+
+internal object ArgTypePath: ArgType<Path>(hasParameter = true) {
+    override val description: kotlin.String get() = "{ String }"
+    override fun convert(value: kotlin.String, name: kotlin.String): Path = value.toPath(normalize = true)
+}
