--- tools/differ/api/differ.api
@@ -1,4 +1,0 @@
-public final class io/matthewnelson/differ/MainKt {
-	public static final fun main ([Ljava/lang/String;)V
-}
-
