--- tools/differ/src/commonMain/kotlin/io/matthewnelson/differ/internal/apply/DirApply.kt
@@ -1,93 +1,0 @@
-/*
- * Copyright (c) 2023 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package io.matthewnelson.differ.internal.apply
-
-import io.matthewnelson.differ.internal.ArgDiffDir
-import io.matthewnelson.differ.internal.OptDiffFileExtName
-import io.matthewnelson.differ.internal.Subcommand
-import io.matthewnelson.differ.internal.create.DirCreate
-import okio.FileSystem
-import okio.Path
-
-internal abstract class DirApply(
-    private val fs: FileSystem,
-    private val runner: Runner,
-): Subcommand(
-    name = NAME_CMD,
-    description = """
-        Applies diff files from previously created
-        ${DirCreate.NAME_CMD} to the specified $NAME_DIR.
-        Files from $NAME_DIR are modified in place.
-    """,
-    additionalIndent = 1,
-),  ArgDiffDir,
-    OptDiffFileExtName
-{
-
-    protected abstract val dirArg: Path
-    abstract override val diffDirArg: Path
-    abstract override val diffFileExtNameOpt: String
-
-    final override fun execute() {
-        // TODO: Validate
-
-        try {
-            runner.run(
-                settings = settings(),
-                fs = fs,
-            )
-        } catch (t: Throwable) {
-            // TODO: Cleanup
-            throw t
-        }
-    }
-
-    internal interface Runner {
-
-        @Throws(Throwable::class)
-        fun run(settings: Settings, fs: FileSystem, )
-
-        companion object: Runner {
-
-            @Throws(Throwable::class)
-            override fun run(settings: Settings, fs: FileSystem, ) {
-                // TODO
-            }
-        }
-    }
-
-    internal companion object {
-        internal const val NAME_CMD = "dir-apply"
-
-        internal const val NAME_DIR = "dir"
-
-        internal fun from(
-            fs: FileSystem,
-            runner: Runner,
-            dir: Path,
-            diffDir: Path,
-            diffFileExtName: String,
-            settings: Settings,
-        ): DirApply {
-            return object : DirApply(fs = fs, runner = runner) {
-                override val dirArg: Path = dir
-                override val diffDirArg: Path = diffDir
-                override val diffFileExtNameOpt: String = diffFileExtName
-                override val quietOpt: Boolean = settings.quiet
-            }
-        }
-    }
-}
